heapslice: Avoid int to iface allocs
Converting an int to an interface incurs an alloc for larger ints
but not for small ints due to an optimization made in go 1.15:
https://github.com/golang/go/commit/9828c43288a53d3df75b1f73edad0d037a91dff8

This optimization only works for 0-255, hence the 100 case which used
0-100 was not impacted.

We can avoid this allocation by skipping use of the interface argument
value for `heap.Push` and `heap.Pop`, and instead rely on a typed field
in the heap wrapper. Since we have to pass/return a value through the
interface, we use 0, which doesn't alloc.